+++ src/gtgt/mutalyzer.py
@@ -477,6 +477,8 @@ class Therapy:
     description: str
     variants: Sequence[Variant]
     figure: OptionalType[str] = None
+    hgvsr: OptionalType[str] = None
+    hgvsp: OptionalType[str] = None
 
     @classmethod
     def from_dict(cls, dict: Mapping[str, Any]) -> "Therapy":
@@ -488,6 +490,8 @@ class Therapy:
             description=dict["description"],
             variants=v,
             figure=dict.get("figure"),
+            hgvsr=dict.get("hgvsr"),
+            hgvsp=dict.get("hgvsp"),
         )
 
 
@@ -695,9 +699,17 @@ def skip_adjacent_exons(d: Description, number_to_skip: int = 1) -> Sequence[The
         name = f"Skip {exons_description}"
         selector = d.get_selector_id()
         cdot_variants = to_cdot_hgvs(d, combined)
-        hgvs = f"{selector}:c.{cdot_variants}"
+        hgvsc = f"{selector}:c.{cdot_variants}"
+        hgvsr = f"{selector}:r.{cdot_variants}"
         description = description
-        t = Therapy(name, hgvs, description, combined)
+        t = Therapy(
+            name=name,
+            hgvsc=hgvsc,
+            hgvsr=hgvsr,
+            hgvsp=protein_prediction(d, combined)[0],
+            description=description,
+            variants=combined,
+        )
         exon_skips.append(t)
 
     return exon_skips