+++ src/gtgt/transcript.py
@@ -13,6 +13,7 @@ from .mutalyzer import (
     generate_therapies,
     init_description,
     mutation_to_cds_effect,
+    protein_prediction,
     sequence_from_description,
 )
 
@@ -167,6 +168,8 @@ class Transcript:
         wt = Therapy(
             name="Wildtype",
             hgvsc=hgvs.split("c.")[0] + "c.=",
+            hgvsr=hgvs.split("c.")[0] + "r.=",
+            hgvsp=protein_prediction(d, [])[0],
             description="These are the annotations as defined on the reference. They are always 100% by definition.",
             variants=list(),
         )
@@ -177,6 +180,8 @@ class Transcript:
         input = Therapy(
             name="Input",
             hgvsc=hgvs,
+            hgvsr=hgvs.replace(":c.", ":r."),
+            hgvsp=protein_prediction(d, input_variants)[0],
             description="The annotations based on the supplied input variants.",
             variants=input_variants,
             figure=draw(d),